Added eventType view and backup.sql

+++ src/bd2/EventType.java
@@ -11,17 +11,17 @@ import org.hibernate.annotations.GenericGenerator;
 @Entity
 @Table(name="typy_wydarzen")
 public class EventType {
-	private byte id;
+	private int id;
 	private String typeName;
 	
 	@Id
 	@Column(name="id_typu")
 	@GeneratedValue(generator="increment")
 	@GenericGenerator(name="increment", strategy="increment")
-	public byte getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(byte id) {
+	public void setId(int id) {
 		this.id = id;
 	}
